backend/requirements.txt
# CORE FRAMEWORK
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# AUTHENTICATION & SECURITY
djangorestframework-simplejwt==5.3.1

# REAL-TIME & WEBSOCKETS (ASGI)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# DATABASE & CACHING
psycopg2==2.9.9
redis==5.0.1
django-redis==5.4.0

# UTILITIES & PRODUCTION SERVERS
python-dotenv==1.0.0
Pillow==10.1.0
gunicorn==21.2.0